cmd/k8s-operator: fix image pull secrets usage for proxy
Set imagePullSecrets for operator and proxy in service accounts instead
of the operator deployment. Without this only the operator can pull the
image if imagePullSecrets value is set, so it is impossible to use the
proxy or create subnet routers in environments where anonymous auth
pulling is not allowed.

The logical place to configure this is in the service accounts for both
operator and proxy, not the operator deployment template.

Fixes #16772

Signed-off-by: Ivan Tarasov <ivan.tarasov@gmail.com>